改进游戏过程中的速度（小偷的速度）使不同性能的设备不会因为性能问题而有拖慢现象 主要指从int到float 因为GameActor的actorX本来就是float的用int的处理再转回来太慢 还有就是把游戏速度定义为（半个屏幕/毫秒）可以有效减少计算次数 之前每次用到游戏速度都要转换成毫秒 这样一次转换大家受益

+++ God'sHandAndThief/src/org/bistu/xgxykjcx/godshandandthief/actor/GameActor.java
@@ -83,9 +83,9 @@ public abstract class GameActor {
 		return false;
 	}
 	
-	abstract public int getLeft();
+	abstract public float getLeft();
 	
-	abstract public int getRight();
+	abstract public float getRight();
 	
 	public int getWidth() {
 		return width;
